Text summarization using YT video with LangChain/requirements.txt
chainlit
openai
langchain
python-dotenv
langchain_openai 
youtube-transcript-api
pytube